% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text_arguments_parser.R
\name{convert_datasets_to_id}
\alias{convert_datasets_to_id}
\title{Convert dataset names to IDs}
\usage{
convert_datasets_to_id(name, use_dev = FALSE, api_key = NULL)
}
\arguments{
\item{name}{Character vector of dataset titles or numeric IDs.}

\item{use_dev}{Logical; if TRUE, use the development version of MDV.}

\item{api_key}{Optional API key for MDV.}
}
\value{
Numeric vector of IDs.
}
\description{
Convert dataset names to IDs
}
\keyword{internal}
