% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{create}
\alias{create}
\title{Create a new object in the catalog}
\usage{
create(object, ...)
}
\arguments{
\item{object}{The object to create (Dataset or Distribution)}
}
\value{
The created object with updated information from the API response
}
\description{
Generic method to create a new object in the data catalog via the API.
}
\keyword{internal}
