% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create.R, R/create_distribution.R
\name{create_distribution}
\alias{create_distribution}
\title{Create a new distribution via API}
\usage{
create_distribution(
  title,
  dataset_id,
  stat_server_flag = NULL,
  zh_web_flag = NULL,
  ogd_flag = NULL,
  description = NULL,
  access_url = NULL,
  byte_size = NULL,
  status_id = NULL,
  license_id = NULL,
  file_format_id = NULL,
  periodicity_id = NULL,
  file_path = NULL,
  start_date = NULL,
  end_date = NULL,
  modified_next = NULL,
  api_key = NULL,
  verbosity = 0,
  use_dev = TRUE,
  preview = FALSE
)
}
\arguments{
\item{title}{Title of the distribution (required; max. 1000 characters).}

\item{dataset_id}{ID of the dataset to which this distribution belongs (required).}

\item{stat_server_flag}{Logical; should the distribution be published on the statistical server? (optional).}

\item{zh_web_flag}{Logical; should the distribution be shown on the Zurich web catalog? (optional).}

\item{ogd_flag}{Logical; should the distribution be published on the OGD portal? (optional).}

\item{description}{Optional free-text description.}

\item{access_url}{Optional URL to access the distribution (must start with http:// or https://).}

\item{byte_size}{Optional file size in bytes (must be a positive number).}

\item{status_id}{Optional character; status ID (will be set via follow-up PATCH).
\code{0} = "verworfen,
\code{1} = "✍️ Entwurf ",
\code{2} = "🔍 in Prüfung",
\code{3} = "📜 publiziert",}

\item{license_id}{integer; license ID.
Use \code{1} = "CC BY 4.0 (Attribution required)" or \code{2} = "CC0 (No attribution required)".}

\item{file_format_id}{Optional file format ID.
Häufige Formate:
\code{3} = "CSV"
\code{20} = "PDF"
\code{33} = "XLSX"
\code{12} = "JSON"
\code{29} = "TXT"
\code{34} = "XML"
\code{32} = "XLS"
\code{35} = "ZIP"
\code{8} = "HTML"
\code{13} = "JSONLD"
\code{5} = "DOCX"
\code{40} = "GPKG"
\code{11} = "JS"
\code{25} = "RTF"}

\item{periodicity_id}{Optional character periodicity ID.
\code{1} = "Jährlich",
\code{21} = "Kontinuierlich",
\code{42} = "Zweijährlich",
\code{43} = "Halbjährlich",
\code{45} = "Vierteljährlich",
\code{47} = "Monatlich",
\code{49} = "Alle zwei Wochen",
\code{51} = "Wöchentlich",
\code{52} = "Halbwöchentlich",
\code{54} = "Täglich",
\code{55} = "Unregelmässig",
\code{56} = "Niemals"}

\item{file_path}{Optional local file path; if provided, the file will be uploaded and linked.}

\item{start_date}{POSIXct or ISO datetime string; new start (optional)}

\item{end_date}{POSIXct or ISO datetime string; new end (optional)}

\item{modified_next}{POSIXct or ISO datetime string; new next update of Dataset (optional)}

\item{api_key}{Optional API key; if not provided, the default environment variable is used.}

\item{verbosity}{Integer; verbosity level passed to httr2::req_perform() (default: 0).}

\item{use_dev}{Logical; whether to use the development API endpoint (default: TRUE).}

\item{preview}{Defines if it is a test run. If TRUE, the
dataset-object is returned into the environment.
Default = FALSE}
}
\value{
Invisibly returns the parsed API response as a list.
}
\description{
Create a new distribution via API
}
\details{
If \code{file_path} is provided, the file will be uploaded via a separate API call prior to creating the distribution.
The resulting \code{file_upload_id} and \code{file_format_id} will be automatically extracted and included in the distribution payload.

If \code{status_id} is specified, it will be applied via a follow-up PATCH request after the distribution is created,
as the initial POST request does not accept a custom status.

If \code{start_date} or \code{end_date} are provided, the dataset's overall timespan will be updated via \code{update_dataset()}.

To assist in constructing valid queries, the following functions provide the
set of acceptable values:
\itemize{
\item \code{get_statuses()}
\item \code{get_licenses()}
\item \code{get_formats()}
\item \code{get_periodicities()}
}

Use these functions to inspect the available values before making a query.
}
