% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_upload.R
\name{FileUpload}
\alias{FileUpload}
\title{FileUpload class}
\usage{
FileUpload(file_path = S7::class_missing)
}
\arguments{
\item{file_path}{Path to the local file to upload (must exist).}
}
\value{
An S7 \code{FileUpload} object.
}
\description{
Defines a new \code{FileUpload} object representing a local file to be uploaded
via the API.
}
