% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{get_base_url}
\alias{get_base_url}
\title{Get the base URL based on environment setting}
\usage{
get_base_url(use_dev = FALSE)
}
\arguments{
\item{use_dev}{Whether to use the development environment}
}
\value{
The base URL string
}
\description{
Get the base URL based on environment setting
}
\keyword{internal}
