% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text_arguments_parser.R
\name{get_datasets}
\alias{get_datasets}
\title{Get All Datasets and Their IDs}
\usage{
get_datasets(input = NULL, use_dev = FALSE, api_key = NULL)
}
\arguments{
\item{input}{Optional character vector of dataset titles or numeric IDs.}

\item{use_dev}{boolean FALSE = developemtent version of MDV}

\item{api_key}{optional API key for MDV}
}
\value{
A tibble with columns:
\itemize{
\item \code{dataset} (character): dataset title
\item \code{id} (numeric): dataset ID
}
}
\description{
Retrieves a tibble of datasets (title and id), with optional filtering.
}
\examples{
\dontrun{
  # All datasets
  get_datasets()

  # Filter by title
  get_datasets("Hotels")

  # Filter by ID
  get_datasets(10)
}

}
