% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text_arguments_parser.R
\name{get_formats}
\alias{get_formats}
\title{Get All Formats and Their IDs}
\usage{
get_formats(input = NULL, use_dev = FALSE, api_key = NULL)
}
\arguments{
\item{use_dev}{boolean FALSE = developemtent version of MDV}

\item{api_key}{optional API key for MDV}
}
\description{
Get All Formats and Their IDs
}
\keyword{internal}
