% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text_arguments_parser.R
\name{get_licenses}
\alias{get_licenses}
\title{Get All Licenses and Their IDs}
\usage{
get_licenses(input = NULL, use_dev = FALSE, api_key = NULL)
}
\arguments{
\item{input}{a set of licenses to retrieve. Empty = all}

\item{use_dev}{boolean FALSE = development version of MDV}

\item{api_key}{optional API key for MDV}
}
\description{
Get All Licenses and Their IDs
}
