% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text_arguments_parser.R
\name{get_licenses_id}
\alias{get_licenses_id}
\title{Retrieve License ID by Name}
\usage{
get_licenses_id(name)
}
\arguments{
\item{name}{A character string or vector specifying the name(s) of the license(s).}
}
\value{
An integer value or vector representing the ID(s) of the license(s).
}
\description{
This function returns the ID associated with a license name by first retrieving
a data frame of licenses and then matching the provided name to its
corresponding ID.
}
\examples{
\dontrun{
get_licenses_id("NonCommercialAllowed-CommercialAllowed-ReferenceRequired")
}
}
