% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{get_media_types}
\alias{get_media_types}
\title{Get available media types for distributions}
\usage{
get_media_types(auth_info, use_dev = TRUE)
}
\arguments{
\item{auth_info}{Authentication information from login_to_api()}

\item{use_dev}{Whether to use the development environment (default: TRUE)}
}
\value{
A list of available media types
}
\description{
Retrieves all available media types that can be used for distributions.
}
