% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text_arguments_parser.R
\name{get_periodicities_id}
\alias{get_periodicities_id}
\title{Retrieve Periodicity ID by Name}
\usage{
get_periodicities_id(name)
}
\arguments{
\item{name}{A character string or vector specifying the name(s) of the
periodicity (or periodicities).}
}
\value{
An integer value or vector representing the ID(s) of the periodicity
(or periodicities).
}
\description{
This function returns the ID associated with a periodicity name by first
retrieving a data frame of periodicities and then matching the provided name
to its corresponding ID.
}
\examples{
get_periodicity_id("Jährlich")
}
