% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text_arguments_parser.R
\name{get_statuses}
\alias{get_statuses}
\title{Get All Statuses and Their IDs}
\usage{
get_statuses(input = NULL, use_dev = FALSE, api_key = NULL)
}
\arguments{
\item{input}{Optional character vector of status names or numeric IDs.}

\item{use_dev}{boolean FALSE = developemtent version of MDV}

\item{api_key}{optional API key for MDV}
}
\value{
A tibble with two columns:
\itemize{
\item \code{status} (character): the status label
\item \code{id} (numeric): the status ID
}
}
\description{
Retrieves a tibble of all statuses and their IDs. Optionally filters by name or ID.
}
\examples{
\dontrun{
  # All statuses
  get_statuses()

  # By name
  get_statuses("Entwurf")

  # By ID
  get_statuses(3)
}
}
