% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{\%<>\%}
\alias{\%<>\%}
\title{Get first or second argument}
\usage{
x \%<>\% y
}
\arguments{
\item{x}{Object 1}

\item{y}{Object 2}
}
\description{
Returns the first argument if it is not \code{NULL}, otherwise returns the second.
}
