% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/set_status.R
\name{set_status}
\alias{set_status}
\title{Set the status of a Dataset or Distribution via the "/set-status" endpoint}
\usage{
set_status(object, ...)
}
\arguments{
\item{object}{A \code{Dataset} or \code{Distribution} S7 object with \verb{@id} and \verb{@status_id} set}
}
\value{
Invisibly returns parsed response as a list.
}
\description{
Set the status of a Dataset or Distribution via the "/set-status" endpoint
}
\keyword{internal}
