% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update.R, R/update_distribution.R
\name{update_distribution}
\alias{update_distribution}
\title{Update an existing distribution via API}
\usage{
update_distribution(
  id,
  title = NULL,
  description = NULL,
  stat_server_flag = NULL,
  zh_web_flag = NULL,
  ogd_flag = NULL,
  access_url = NULL,
  byte_size = NULL,
  status_id = NULL,
  license_id = NULL,
  file_format_id = NULL,
  dataset_id = NULL,
  periodicity_id = NULL,
  file_path = NULL,
  start_date = NULL,
  end_date = NULL,
  modified_next = NULL,
  api_key = NULL,
  verbosity = 0,
  use_dev = TRUE,
  preview = FALSE
)
}
\arguments{
\item{id}{ID of the distribution to update (required).}

\item{title}{Title of the distribution (optional; max. 1000 characters).}

\item{description}{Optional free-text description.}

\item{stat_server_flag}{Logical; should the distribution be published on the statistical server? (optional).}

\item{zh_web_flag}{Logical; should the distribution be shown on the Zurich web catalog? (optional).}

\item{ogd_flag}{Logical; should the distribution be published on the OGD portal? (optional).}

\item{access_url}{Optional URL to access the distribution (must start with http:// or https://).}

\item{byte_size}{Optional file size in bytes (must be a positive number).}

\item{status_id}{Optional status ID (applied via PATCH after update).
\code{0} = "verworfen,
\code{1} = "✍️ Entwurf ",
\code{2} = "🔍 in Prüfung",
\code{3} = "📜 publiziert",}

\item{license_id}{Optional license ID.
Use \code{1} = "CC BY 4.0 (Attribution required)" or \code{2} = "CC0 (No attribution required)".}

\item{file_format_id}{Optional file format ID.
Häufige Formate:
\code{3} = "CSV"
\code{20} = "PDF"
\code{33} = "XLSX"
\code{12} = "JSON"
\code{29} = "TXT"
\code{34} = "XML"
\code{32} = "XLS"
\code{35} = "ZIP"
\code{8} = "HTML"
\code{13} = "JSONLD"
\code{5} = "DOCX"
\code{40} = "GPKG"
\code{11} = "JS"
\code{25} = "RTF"}

\item{dataset_id}{ID of the dataset to which this distribution belongs (optional).}

\item{periodicity_id}{Optional update frequency ID.
\code{1} = "Jährlich",
\code{21} = "Kontinuierlich",
\code{42} = "Zweijährlich",
\code{43} = "Halbjährlich",
\code{45} = "Vierteljährlich",
\code{47} = "Monatlich",
\code{49} = "Alle zwei Wochen",
\code{51} = "Wöchentlich",
\code{52} = "Halbwöchentlich",
\code{54} = "Täglich",
\code{55} = "Unregelmässig",
\code{56} = "Niemals"}

\item{file_path}{Optional local file path; if provided, the file will be uploaded and linked.}

\item{start_date}{POSIXct or ISO datetime string; new start of Dataset(optional)}

\item{end_date}{POSIXct or ISO datetime string; new end of Dataset (optional)}

\item{modified_next}{POSIXct or ISO datetime string; new next update of Dataset (optional)}

\item{api_key}{Optional API key; if not provided, the default environment variable is used.}

\item{verbosity}{Integer; verbosity level passed to httr2::req_perform() (default: 0).}

\item{use_dev}{Logical; whether to use the development API endpoint (default: TRUE).}

\item{preview}{Defines if it is a test run. If TRUE, the
dataset-object is returned into the environment.
Default = FALSE}
}
\value{
Invisibly returns the parsed API response as a list.
}
\description{
Update an existing distribution via API
}
\details{
If \code{file_path} is provided, the file will be uploaded via a separate API call prior to updating the distribution.
The resulting \code{file_upload_id} and \code{file_format_id} will be automatically extracted and included in the update payload.

If \code{status_id} is specified, it will be applied via a separate PATCH request after the distribution is updated.

If \code{start_date} or \code{end_date} are provided, the dataset's overall timespan will be updated via \code{update_dataset()}.
}
