% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{api_request}
\alias{api_request}
\title{Helper for API calls}
\usage{
api_request(
  method = c("GET", "POST", "PUT", "PATCH", "DELETE"),
  endpoint,
  payload = NULL,
  api_key,
  use_dev = TRUE
)
}
\description{
Helper for API calls
}
