% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{api_request}
\alias{api_request}
\title{Helper for API calls}
\usage{
api_request(
  method = c("GET", "POST", "PUT", "PATCH", "DELETE"),
  endpoint,
  payload = NULL,
  api_key,
  use_dev = TRUE
)
}
\arguments{
\item{method}{string; what kind of request should be made}

\item{endpoint}{string; which endpoint should be reached}

\item{payload}{list containing the payload}

\item{api_key}{MDV API key (optional; falls back to env var).}

\item{use_dev}{Logical; if TRUE, uses the development API endpoint.}
}
\description{
Helper for API calls
}
\keyword{internal}
