% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text_arguments_parser.R
\name{convert_datasets_to_id}
\alias{convert_datasets_to_id}
\title{Retrieve Dataset ID by Name}
\usage{
convert_datasets_to_id(name)
}
\arguments{
\item{name}{A character string or vector specifying the name(s) of the dataset(s).}
}
\value{
An integer value or vector representing the ID(s) of the dataset(s).
}
\description{
This function returns the ID associated with a dataset by first retrieving a
data frame of dataset-names and then matching the provided name to its
corresponding ID.
}
\examples{
\dontrun{
convert_keywords_to_id("abfall")
}
}
