% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text_arguments_parser.R
\name{convert_keywords_to_id}
\alias{convert_keywords_to_id}
\title{Retrieve Keyword ID by Name}
\usage{
convert_keywords_to_id(name)
}
\arguments{
\item{name}{A character string or vector specifying the name(s) of the keyword(s).}
}
\value{
An integer value or vector representing the ID(s) of the keyword(s).
}
\description{
This function returns the ID associated with a keyword by first retrieving a
data frame of keywords and then matching the provided name to its
corresponding ID.
}
\examples{
\dontrun{
convert_keywords_to_id("abfall")
}
}
