% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text_arguments_parser.R
\name{convert_keywords_to_id}
\alias{convert_keywords_to_id}
\title{Convert keyword names to IDs}
\usage{
convert_keywords_to_id(name)
}
\arguments{
\item{name}{Character vector of keyword names.}
}
\value{
Numeric vector of IDs.
}
\description{
Convert keyword names to IDs
}
\keyword{internal}
