% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text_arguments_parser.R
\name{convert_licenses_to_id}
\alias{convert_licenses_to_id}
\title{Convert license names to IDs}
\usage{
convert_licenses_to_id(name)
}
\description{
Convert license names to IDs
}
\keyword{internal}
