% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text_arguments_parser.R
\name{convert_periodicities_to_id}
\alias{convert_periodicities_to_id}
\title{Convert periodicity names to IDs}
\usage{
convert_periodicities_to_id(name)
}
\description{
Convert periodicity names to IDs
}
\keyword{internal}
