% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text_arguments_parser.R
\name{convert_statuses_to_id}
\alias{convert_statuses_to_id}
\title{Convert status names to IDs}
\usage{
convert_statuses_to_id(name, use_dev = FALSE, api_key = NULL)
}
\arguments{
\item{use_dev}{boolean FALSE = developemtent version of MDV}

\item{api_key}{optional API key for MDV}
}
\description{
Convert status names to IDs
}
\keyword{internal}
