% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text_arguments_parser.R
\name{convert_themes_to_id}
\alias{convert_themes_to_id}
\title{Convert theme names to IDs}
\usage{
convert_themes_to_id(name)
}
\description{
Convert theme names to IDs
}
\keyword{internal}
\keyword{keywords}
