% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text_arguments_parser.R
\name{convert_zh_web_catalog_to_id}
\alias{convert_zh_web_catalog_to_id}
\title{Retrieve zh-web-catalog keyword ID by Name}
\usage{
convert_zh_web_catalog_to_id(name)
}
\arguments{
\item{name}{A character string or vector specifying the name(s) of the
zh-web-catalog keyword(s).}
}
\value{
An integer value or vector representing the ID(s) of the
zh-web-catalog keyword(s)
}
\description{
This function returns the ID associated with a zh-web-catalog keyword by
first retrieving a data frame of keywords and then matching the provided name
to its corresponding ID.
}
\examples{
\dontrun{
convert_zh_web_catalog_to_id("Bevölkerung")
}
}
