% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text_arguments_parser.R
\name{convert_zh_web_catalog_to_id}
\alias{convert_zh_web_catalog_to_id}
\title{Convert zh-web-catalog names to IDs}
\usage{
convert_zh_web_catalog_to_id(name)
}
\description{
Convert zh-web-catalog names to IDs
}
\keyword{internal}
\keyword{keywords}
