% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{create}
\alias{create}
\title{Create a new object in the catalog}
\usage{
create(object, ...)
}
\arguments{
\item{object}{The object to create (Dataset or Distribution)}

\item{auth_info}{Authentication information from login_to_api()}

\item{use_dev}{Whether to use the development environment}
}
\value{
The created object with updated information from the API response
}
\description{
Generic method to create a new object in the data catalog via the API.
}
\keyword{internal}
