% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_distribution.R
\name{Distribution}
\alias{Distribution}
\title{Distribution class for DCAT API}
\usage{
Distribution(
  title = class_missing,
  description = class_missing,
  dataset_id = class_missing,
  access_url_id = class_missing,
  byte_size = class_missing,
  download_url_id = class_missing,
  format_id = class_missing,
  identifier = class_missing,
  issued = class_missing,
  media_type_id = class_missing,
  ogd_flag = class_missing,
  right_id = class_missing,
  sort_order = class_missing,
  stat_server_flag = class_missing,
  status_id = class_missing,
  zhweb_flag = class_missing
)
}
\description{
An S7 class representing a DCAT distribution (file or access point)
associated with a dataset in the catalog.
}
\details{
This class implements the DCAT AP CH V2 standard for distribution metadata.
}
\keyword{internal}
