% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distribution.R
\name{Distribution}
\alias{Distribution}
\title{Distribution}
\usage{
Distribution(
  id = S7::class_missing,
  title = S7::class_missing,
  dataset_id = S7::class_missing,
  stat_server_flag = S7::class_missing,
  zh_web_flag = S7::class_missing,
  ogd_flag = S7::class_missing,
  description = S7::class_missing,
  access_url = S7::class_missing,
  byte_size = S7::class_missing,
  status_id = S7::class_missing,
  license_id = S7::class_missing,
  file_format_id = S7::class_missing,
  periodicity_id = S7::class_missing,
  file_upload_id = S7::class_missing
)
}
\arguments{
\item{id}{numeric; ID der Distribution (required for update)}

\item{title}{character; distribution title (required on create, <=1000 characters)}

\item{dataset_id}{numeric; ID of the related dataset (required)}

\item{stat_server_flag}{logical; visibility flag for statistical server (optional)}

\item{zh_web_flag}{logical; visibility flag for zh web portal (optional)}

\item{ogd_flag}{logical; visibility flag for Open Government Data (optional)}

\item{description}{character; distribution description (optional, <=4000 characters)}

\item{access_url}{character; URL to access the distribution (optional, must start with http:// or https://)}

\item{byte_size}{numeric; size in bytes (optional, must be a positive integer)}

\item{status_id}{character; status ID (optional)}

\item{license_id}{numeric; license ID (optional)}

\item{file_format_id}{character; file format ID (optional)}

\item{periodicity_id}{character; periodicity ID (optional)}

\item{file_upload_id}{character; ID of the uploaded file (optional)}
}
\value{
An S7 \code{Distribution} object.
}
\description{
Create a new Distribution object for the KOSMOS data catalog (DCAT standard).
}
