% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{get_api_key}
\alias{get_api_key}
\title{Retrieve MDV API key}
\usage{
get_api_key(key = NULL)
}
\arguments{
\item{key}{Optional plain API key.}
}
\value{
API key string.
}
\description{
Attempts to fetch the API key from (in order):
\enumerate{
\item Provided argument
\item Environment variable \code{MDV_API_KEY}
\item Interactive prompt (RStudio or askpass)
}
}
