% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{get_datasets}
\alias{get_datasets}
\title{Get a list of datasets from the catalog}
\usage{
get_datasets(auth_info, use_dev = TRUE)
}
\arguments{
\item{auth_info}{Authentication information from login_to_api()}

\item{use_dev}{Whether to use the development environment (default: TRUE)}
}
\value{
A list of datasets as returned by the API
}
\description{
Retrieves all datasets available in the data catalog that the user has access to.
}
\examples{
\dontrun{
# First login
auth_info <- login_to_api("admin@example.com", "admin")

# Get all datasets
datasets <- get_datasets(auth_info)
}
}
