% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text_arguments_parser.R
\name{get_datasets}
\alias{get_datasets}
\title{Get All Datasets and Their IDs}
\usage{
get_datasets(input = NULL)
}
\arguments{
\item{input}{Optional. A character string, numeric value, or a vector of either.
Used to filter the "dataset" dataset. If NULL (default), returns the full dataset.}
}
\value{
A data frame with two columns: one for dataset titles and one for
their corresponding IDs.
}
\description{
Retrieves a data frame containing all available datasets along with their
associated IDs from the API. Optionally, it can filter the results based on
an input vector.
}
\examples{
\dontrun{
datasets_df <- get_datasets()
head(datasets_df)

# With input conversion/filtering
get_datasets(10)

get_datasets("Hotels")
}#'
}
