% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{get_formats}
\alias{get_formats}
\title{Get available formats for distributions}
\usage{
get_formats(auth_info, use_dev = TRUE)
}
\arguments{
\item{auth_info}{Authentication information from login_to_api()}

\item{use_dev}{Whether to use the development environment (default: TRUE)}
}
\value{
A list of available formats
}
\description{
Retrieves all available formats that can be used for distributions.
}
