% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text_arguments_parser.R
\name{get_keywords}
\alias{get_keywords}
\title{Get All Keywords and Their IDs}
\usage{
get_keywords(input = NULL)
}
\arguments{
\item{input}{Optional. A character string, numeric value, or a vector of either.
Used to filter the keyword dataset. If NULL (default), returns the full dataset.}
}
\value{
A data frame with two columns: one for keyword names and one for
their corresponding IDs.
}
\description{
Retrieves a data frame containing all available keywords along with their
associated IDs. This function is typically used to look up valid keyword
options for other functions.
}
\examples{
\dontrun{
df_keywords <- get_keywords()
head(df_keywords)

get_keywords("abwasser")

get_keyords(c(578, 590))

}
}
