% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text_arguments_parser.R
\name{get_organisations}
\alias{get_organisations}
\title{Get All Organisations and Their IDs}
\usage{
get_organisations(
  show_organisation_units = TRUE,
  use_dev = FALSE,
  api_key = NULL
)
}
\arguments{
\item{show_organisation_units}{Logical; if TRUE, include sub-units.}

\item{use_dev}{boolean FALSE = development version of MDV}

\item{api_key}{optional API key for MDV}
}
\value{
A tibble with columns:
\itemize{
\item \code{organisation_id} (numeric)
\item \code{organisation} (character)
\item optionally \code{organisation_unit_id} (numeric) and \code{organisation_unit} (character)
}
}
\description{
Retrieves a tibble of all organisations and their IDs. Optionally includes sub-units.
}
\examples{
\dontrun{
  # All organisations
  get_organisations()

  # Without sub-units
  get_organisations(FALSE)
}
}
