% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text_arguments_parser.R
\name{get_organisations}
\alias{get_organisations}
\title{Get All Organisation and Their IDs}
\usage{
get_organisations(show_organisation_units = TRUE)
}
\arguments{
\item{show_organisation_units}{If \code{TRUE}, organisation units are returned in
addition to the organisations}
}
\value{
A data frame with two columns: one for organisation names and one for
their corresponding IDs.
}
\description{
Retrieves a data frame containing all available organisation along with their
associated IDs. This function is typically used to look up valid organisation
options for other functions.
}
\examples{
df_organisation <- get_organisations()
head(df_organisation)
}
