% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text_arguments_parser.R
\name{get_periodicities}
\alias{get_periodicities}
\title{Get All Periodicities and Their IDs}
\usage{
get_periodicities(input = NULL, use_dev = FALSE, api_key = NULL)
}
\arguments{
\item{input}{Optional character vector of periodicity names or numeric IDs.}

\item{use_dev}{boolean FALSE = developemtent version of MDV}

\item{api_key}{optional API key for MDV}
}
\value{
A tibble with two columns:
\itemize{
\item \code{periodicity} (character): the periodicity label
\item \code{id} (numeric): the periodicity ID
}
}
\description{
Retrieves a tibble of all periodicities and their IDs. Optionally filters by name or ID.
}
\examples{
\dontrun{
  # All periodicities
  get_periodicities()

  # By name
  get_periodicities("Jährlich")

  # By ID
  get_periodicities(42)
}
}
