% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text_arguments_parser.R
\name{req_to_df}
\alias{req_to_df}
\title{Retrieve a Data Frame from API Endpoint Request}
\usage{
req_to_df(endpoint)
}
\arguments{
\item{endpoint}{A character string of a Kosmos API endpoint.}
}
\value{
A data frame with two columns: one for endpoint entries and one for
their corresponding IDs.
}
\description{
For a given Endpoint, this function retrieves the ID's and corresponding
Endpoint's entries and returns it as a data frame.
}
\examples{
df_formats  <- req_to_df("formats")
head(df_formats)
}
