% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text_arguments_parser.R
\name{req_to_df}
\alias{req_to_df}
\title{Retrieve a Data Frame from API Endpoint}
\usage{
req_to_df(endpoint)
}
\arguments{
\item{endpoint}{One of: "keywords", "themes", etc.}
}
\value{
A tibble with columns \verb{<endpoint>} and \code{id}.

A tibble with columns \verb{<endpoint>} and \code{id}.
}
\description{
Generic helper to fetch \if{html}{\out{<endpoint>}} entries with labels and ids.
}
\keyword{<endpoint>}
\keyword{API}
\keyword{Data}
\keyword{Endpoint}
\keyword{Frame}
\keyword{Generic}
\keyword{Retrieve}
\keyword{a}
\keyword{and}
\keyword{entries}
\keyword{fetch}
\keyword{from}
\keyword{helper}
\keyword{ids.}
\keyword{internal}
\keyword{labels}
\keyword{to}
\keyword{with}
