% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{update}
\alias{update}
\title{Update an existing object in the catalog}
\usage{
update(object, ...)
}
\arguments{
\item{object}{The object to update}

\item{id}{ID of the object to update}

\item{auth_info}{Authentication information from login_to_api()}

\item{use_dev}{Whether to use the development environment}
}
\value{
The updated object with updated information from the API response
}
\description{
Generic method to update an existing object in the data catalog via the API.
}
\keyword{internal}
